refactor: MemoirSummaryDto 기본키 타입을 Long으로 변경 - #20 - MemoirSummaryDto의 memoriId 타입을 Long으로 변경했습니다.

+++ backend/open-mission/src/main/java/woowatech/open_mission/DTO/MemoirSummaryDto.java
@@ -2,6 +2,4 @@ package woowatech.open_mission.DTO;
 
 import java.util.Date;
 
-public record MemoirSummaryDto(Integer memoriId, String title, Date date) {
-
-}
+public record MemoirSummaryDto(Long memoriId, String title, Date date) {}
